:recycle: move SequenceProxy to private module
