+++ discord/utils/private.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import array
 import asyncio
+import collections.abc
 import datetime
 import functools
 import re
@@ -26,8 +27,12 @@ from typing import (
     Coroutine,
     Awaitable,
     reveal_type,
+    Generic,
+    Sequence,
+    Iterator,
 )
 
+from . import T_co
 from ..errors import InvalidArgument, HTTPException
 
 if TYPE_CHECKING:
@@ -456,3 +461,31 @@ def copy_doc(original: Callable) -> Callable[[T], T]:
         return overridden
 
     return decorator
+
+
+class SequenceProxy(collections.abc.Sequence, Generic[T_co]):
+    """Read-only proxy of a Sequence."""
+
+    def __init__(self, proxied: Sequence[T_co]):
+        self.__proxied = proxied
+
+    def __getitem__(self, idx: int) -> T_co:
+        return self.__proxied[idx]
+
+    def __len__(self) -> int:
+        return len(self.__proxied)
+
+    def __contains__(self, item: Any) -> bool:
+        return item in self.__proxied
+
+    def __iter__(self) -> Iterator[T_co]:
+        return iter(self.__proxied)
+
+    def __reversed__(self) -> Iterator[T_co]:
+        return reversed(self.__proxied)
+
+    def index(self, value: Any, *args, **kwargs) -> int:
+        return self.__proxied.index(value, *args, **kwargs)
+
+    def count(self, value: Any) -> int:
+        return self.__proxied.count(value)
